+++ odin/resources/resolve.ico
