+++ odin/softwareList/mari.bat
