+++ odin/softwareList/photoshop.bat
