+++ odin/softwareList/resolve.bat
