+++ odin/source/core/assets.py
@@ -1,36 +1,37 @@
 import os
 
-from Odin.source.core import trees_path
-from Odin.source.core.create_tree import Tree
+from . import trees_path
+from .create_tree import Tree
 
 from CommonTools.os_ import make_dirs
 from CommonTools.concat import concat
 
 
-def create_set(root, project, set_name, asset_type="SET"):
+def create_asset(root, project, asset_name, asset_type):
     """
 
     Args:
         root (str): root path of the project without the slash at the end
         project (str): project
-        set_name (str): asset name
-        asset_type (str): SET by default
+        asset_name (str): asset name
+        asset_type (str): CHARA or PROPS
 
     Returns:
         bool: True if the project was created, False if it was not
 
     """
-    set_path = concat(root, project, "DATA/LIB", asset_type.upper(), set_name, separator="/")
-    set_tree = Tree.create_from_template(trees_path.set_tree(), set_path)
+    asset_path = concat(root, project, "DATA/LIB", asset_type.upper(), asset_name, separator="/")
+    asset_tree = Tree.create_from_template(trees_path.asset_tree(), asset_path)
 
-    asset_publish_path = concat(root, project, "DATA/LIB/PUBLISH", asset_type.upper(), set_name, separator="/")
-    asset_publish_tree = Tree.create_from_template(trees_path.set_publish_tree(), asset_publish_path)
+    asset_publish_path = concat(root, project, "DATA/LIB/PUBLISH",
+                                asset_type.upper(), asset_name, separator="/")
+    asset_publish_tree = Tree.create_from_template(trees_path.asset_publish_tree(), asset_publish_path)
 
-    asset_created = make_dirs(set_path)
+    asset_created = make_dirs(asset_path)
     asset_publish_created = make_dirs(asset_publish_path)
 
     if asset_created and asset_publish_created:
-        set_tree.create_on_disk()
+        asset_tree.create_on_disk()
         asset_publish_tree.create_on_disk()
 
         return True
@@ -38,13 +39,13 @@ def create_set(root, project, set_name, asset_type="SET"):
         return False
 
 
-def find_sets(root, project, type_="SET"):
+def find_assets(root, project, type_):
     """
 
     Args:
         root (str):
         project (str):
-        type_ (str): SET folder by default
+        type_ (str): CHARACTER, PROPS folder
 
     Returns:
         list(str): assets found in the folder
