+++ odin/source/globals.py
@@ -1,0 +1,4 @@
+from .logger import log
+
+
+Logger = log("Odin")
