+++ odin/trees/asset_publish.yaml
