--- source/globals.py
@@ -1,4 +1,0 @@
-from Odin.source.logger import log
-
-
-Logger = log("Odin")
